+++ context/app/routes_main.py
@@ -1,4 +1,5 @@
-from flask import Blueprint, render_template, abort, current_app, session, flash
+from flask import (Blueprint, render_template, abort, current_app,
+                   session, flash, request, get_flashed_messages)
 
 from yaml import safe_load as load_yaml
 
@@ -54,9 +55,31 @@ def details(type, uuid):
         type_schema = load_yaml(type_schema_file)
     for_each_validation_error(entity, type_schema, flash)
 
-    details_html = object_as_html(entity)
     provenance = client.get_provenance(uuid)
+    flashed_messages = []
+    errors = get_flashed_messages()
 
+    for error in errors:
+        # Traceback trim is a quick fix https://github.com/hubmapconsortium/portal-ui/issues/145.
+        flashed_messages.append({'message': error.message,
+                                 'issue_url': error.issue_url,
+                                 'traceback': error.__str__()[0:1500]})
+
+    if 'react' in request.args:
+        template = f'pages/details/details_react.html'
+        props = {
+            'flashed_messages': flashed_messages,
+            'entity': entity,
+            'provenance': provenance,
+            'vitessce_conf': client.get_vitessce_conf(),
+        }
+        return render_template(
+            template, type=type, uuid=uuid,
+            title_text='TODO: title_text',
+            flask_data=props
+        )
+
+    details_html = object_as_html(entity)
     if type in {'file'}:  # TODO: As we have other specializations, add them here.
         template = f'pages/details/details_{type}.html'
     else: