+++ context/app/static/js/index.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+
+
+ReactDOM.render( // eslint-disable-next-line no-undef
+  <App flaskData={flaskData} />,
+  document.getElementById('react-content'),
+);