+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.0.11 - In progress
 ### Added
 - Hit Elasticsearch instance and start configuring facets.
+- Add react, webpack, eslint, and styling dependencies.
 - Make the Elasticsearch endpoint part of the config.
 
 ## [v0.0.10](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.10) - 2020/03/06